requirements.txt
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
filelock==3.12.0
fonttools==4.39.4
gensim==4.3.1
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
pydantic==1.10.8
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
-e git+https://github.com/rw422scarlet/opinion_modeling.git@9012da758f8a1e7c512378a187a13e2182cbccb2#egg=src
srsly==2.4.6
sympy==1.12
thinc==8.1.10
torch==2.0.1
torchdata==0.6.1
torchtext==0.15.2
tqdm==4.65.0
tweet-preprocessor==0.6.0
typer==0.7.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
wasabi==1.1.1
zipp==3.15.0
